Add metadata to download buttons, refactor Platform.php, Os.php
Minor formatting changes

Minor formatting changes (again)

Minor formatting changes (again 2)
