+++ lib/Os.php
@@ -1,0 +1,10 @@
+<?php
+namespace LMMS;
+
+enum Os: string
+{
+	case Unknown = "Unknown";
+	case Linux = "Linux" ;
+	case MacOS = "macOS";
+	case Windows = "Windows";
+}
